Separate hasFlash into compatibility.js.
